tl_detector: Basic function implementations for tl detection
This commit I implemented some basic functions in the tl_detector
node. They include a function to output images which can be used
for debug or training set. Also the functions to find the nearest
stop line and traffic light and the light state. This commit will
print out the car, stop line, and traffic light position, and
also the traffic light state.

Hope this commit can serve as a starting point for the classification
part of the traffic light.